crawler/requirements.txt
requests
beautifulsoup4
boto3
google-genai
PyGitHub
python-dotenv
gtrending
tenacity
